update var files and remove elastic7
