+++ config/elastic.schema.product.json
@@ -82,11 +82,11 @@
                 "id": {"type": "integer"},
                 "attribute_id": {"type": "integer"},
                 "default_frontend_label": {"type": "text"},
-                "is_visible_on_front": {"type": "boolean"},
+                "is_visible_on_front": {"type": "text"},
                 "is_visible" : {"type": "boolean"},
                 "frontend_input": {"type": "text"},
                 "is_user_defined": {"type": "boolean"},
-                "is_comparable": {"type": "boolean"},
+                "is_comparable": {"type": "text"},
                 "attribute_code": {"type": "text"},
                 "options": {
                     "properties": {
