+++ package.json
@@ -19,18 +19,12 @@
     "cache": "node ./scripts/cache",
     "mage2vs": "node scripts/mage2vs.js",
     "restore": "node scripts/elastic.js restore && npm run db rebuild",
-    "restore7": "node scripts/elastic7.js restore --output-index=vue_storefront_catalog && npm run db rebuild --indexName=vue_storefront_catalog",
-    "restore7_it": "npm run restore7 -- --input-file=var/catalog_it.json --output-index=vue_storefront_catalog_it && npm run db rebuild -- --indexName=vue_storefront_catalog_it",
-    "restore7_de": "npm run restore7 -- --input-file=var/catalog_de.json --output-index=vue_storefront_catalog_de && npm run db rebuild -- --indexName=vue_storefront_catalog_de",
     "restore_it": "npm run restore -- --input-file=var/catalog_it.json --output-index=vue_storefront_catalog_it && npm run db rebuild -- --indexName=vue_storefront_catalog_it",
     "restore_de": "npm run restore -- --input-file=var/catalog_de.json --output-index=vue_storefront_catalog_de && npm run db rebuild -- --indexName=vue_storefront_catalog_de",
     "restore2main": "npm run restore -- --output-index=vue_storefront_catalog",
     "dump": "node scripts/elastic.js dump",
-    "dump_it": "npm run dump -- --input-index=vue_storefront_catalog_it --output-file=var/catalog_it.json",
-    "dump_de": "npm run dump -- --input-index=vue_storefront_catalog_de --output-file=var/catalog_de.json",
-    "dump7": "node scripts/elastic7.js dump",
-    "dump7_it": "npm run dump7 -- --input-index=vue_storefront_catalog_it --output-file=var/catalog_it.json",
-    "dump7_de": "npm run dump7 -- --input-index=vue_storefront_catalog_de --output-file=var/catalog_de.json",
+    "dump_it": "npm run dump -- --input-index=vue_storefront_catalog --output-file=var/catalog_it.json",
+    "dump_de": "npm run dump -- --input-index=vue_storefront_catalog --output-file=var/catalog_de.json",
     "kue": "node scripts/kue.js",
     "kue-dashboard": "npm run kue dashboard -- --port=3050",
     "o2m": "node src/worker/order_to_magento2.js start",
