+++ var/catalog_it_taxrule.json
@@ -1,1 +1,1 @@
-{"_index":"vue_storefront_catalog_it_taxrule","_type":"_doc","_id":"2","_score":1,"_source":{"id":2,"code":"General Taxes","priority":0,"position":0,"customer_tax_class_ids":[3],"product_tax_class_ids":[2],"tax_rate_ids":[6,5,7,4],"calculate_subtotal":false,"rates":[{"id":5,"tax_country_id":"IT","tax_region_id":0,"tax_postcode":"*","rate":23,"code":"VAT23-IT","titles":[]},{"id":4,"tax_country_id":"US","tax_region_id":0,"tax_postcode":"*","rate":23,"code":"VAT23","titles":[]},{"id":6,"tax_country_id":"DE","tax_region_id":0,"tax_postcode":"*","rate":23,"code":"VAT23-DE","titles":[]},{"id":7,"tax_country_id":"PL","tax_region_id":0,"tax_postcode":"*","rate":23,"code":"VAT23-PL","titles":[]}],"tsk":1568837763644}}
+{"_index":"vue_storefront_catalog_taxrule_1587031505","_type":"_doc","_id":"2","_score":1,"_source":{"id":2,"code":"General Taxes","priority":0,"position":0,"customer_tax_class_ids":[3],"product_tax_class_ids":[2],"tax_rate_ids":[6,5,7,4],"calculate_subtotal":false,"rates":[{"id":6,"tax_country_id":"DE","tax_region_id":0,"tax_postcode":"*","rate":23,"code":"VAT23-DE","titles":[]},{"id":4,"tax_country_id":"US","tax_region_id":0,"tax_postcode":"*","rate":23,"code":"VAT23","titles":[]},{"id":5,"tax_country_id":"IT","tax_region_id":0,"tax_postcode":"*","rate":23,"code":"VAT23-IT","titles":[]},{"id":7,"tax_country_id":"PL","tax_region_id":0,"tax_postcode":"*","rate":23,"code":"VAT23-PL","titles":[]}],"tsk":1587030446469}}
