+++ var/catalog_taxrule.json
@@ -1,1 +1,1 @@
-{"_index":"vue_storefront_catalog_taxrule_1568838132","_type":"_doc","_id":"2","_score":1,"_source":{"id":2,"code":"General Taxes","priority":0,"position":0,"customer_tax_class_ids":[3],"product_tax_class_ids":[2],"tax_rate_ids":[6,5,7,4],"calculate_subtotal":false,"rates":[{"tax_region_id":0,"code":"VAT23","rate":23,"id":4,"tax_country_id":"US","titles":[],"tax_postcode":"*"},{"tax_region_id":0,"code":"VAT23-DE","rate":23,"id":6,"tax_country_id":"DE","titles":[],"tax_postcode":"*"},{"tax_region_id":0,"code":"VAT23-PL","rate":23,"id":7,"tax_country_id":"PL","titles":[],"tax_postcode":"*"},{"tax_region_id":0,"code":"VAT23-IT","rate":23,"id":5,"tax_country_id":"IT","titles":[],"tax_postcode":"*"}],"tsk":1568838056364}}
+{"_index":"vue_storefront_catalog_taxrule_1587031505","_type":"_doc","_id":"2","_score":1,"_source":{"id":2,"code":"General Taxes","priority":0,"position":0,"customer_tax_class_ids":[3],"product_tax_class_ids":[2],"tax_rate_ids":[6,5,7,4],"calculate_subtotal":false,"rates":[{"id":6,"tax_country_id":"DE","tax_region_id":0,"tax_postcode":"*","rate":23,"code":"VAT23-DE","titles":[]},{"id":4,"tax_country_id":"US","tax_region_id":0,"tax_postcode":"*","rate":23,"code":"VAT23","titles":[]},{"id":5,"tax_country_id":"IT","tax_region_id":0,"tax_postcode":"*","rate":23,"code":"VAT23-IT","titles":[]},{"id":7,"tax_country_id":"PL","tax_region_id":0,"tax_postcode":"*","rate":23,"code":"VAT23-PL","titles":[]}],"tsk":1587030446469}}
